+++ src/Pagination/PaginationContext.php
@@ -23,7 +23,7 @@ final class PaginationContext
      * @param string $nextUrlPattern URL pattern for next page links. Must contain {@see URL_PLACEHOLDER}.
      * @param string $previousUrlPattern URL pattern for previous page links. Must contain {@see URL_PLACEHOLDER}.
      * @param string $firstPageUrl URL used on the first page.
-     * @param bool $enableAccessibility Whether pagination widgets should add accessibility attributes `aria-*`
+     * @param bool $accessibility Whether pagination widgets should add accessibility attributes `aria-*`
      * automatically.
      * @param TranslatorInterface|null $translator Translator used for pagination messages.
      * @param string $translationCategory Category used with the translator.
@@ -32,7 +32,7 @@ final class PaginationContext
         public readonly string $nextUrlPattern,
         public readonly string $previousUrlPattern,
         public readonly string $firstPageUrl,
-        public readonly bool $enableAccessibility = false,
+        public readonly bool $accessibility = false,
         private readonly ?TranslatorInterface $translator = null,
         private readonly string $translationCategory = BaseListView::DEFAULT_TRANSLATION_CATEGORY,
     ) {}